Rename use-sidebar-workspaces-list.pure.ts to sidebar-workspaces-view-model.ts
