+++ packages/app/src/hooks/sidebar-workspaces-view-model.ts
